feat: setup i18n with next-intl
