+++ src/app/[locale]/(protected)/layout.tsx
