+++ src/app/[locale]/auth/layout.tsx
