+++ src/app/[locale]/auth/login/login-hero.tsx
