--- src/components/NavBar/index.ts
@@ -1,3 +1,0 @@
-export * from './data'
-export { default } from './NavBar'
-export * from './types'
