+++ src/components/layout/header/index.tsx
@@ -1,172 +1,182 @@
 'use client'
+import { User } from 'lucide-react'
+import { useEffect, useRef, useState } from 'react'
 
-import { Badge } from '@heroui/badge'
-import {
-  Calendar,
-  Coffee,
-  Flame,
-  Gift,
-  ShoppingCart,
-  User,
-  UserRound,
-} from 'lucide-react'
-import Image from 'next/image'
-import Link from 'next/link'
+import AccountDropdown from '@/components/layout/header/account-dropdown'
+import LocaleSwitcher from '@/components/layout/header/locale-switcher'
+import MobileMenu from '@/components/layout/header/mobile-menu'
 
-import { useIsMobile } from '@/hooks/use-media-query'
+export interface NavItem {
+  label: string
+  href?: string
+}
 
-const NavBar = () => {
-  const isMobile = useIsMobile()
+export interface NavBarProps {
+  logo: string
+  navItems: NavItem[]
+  cartItemCount: number
+  bgColor?: string
+  textColor?: string
+  themeColor?: string
+  buttonTextColor?: string
+}
 
-  const leftNavItems = [
-    {
-      name: 'trang chủ',
-      href: '/',
-    },
-    {
-      name: 'menu',
-      href: '/menu',
-    },
-    {
-      name: 'phần thưởng',
-      href: '/rewards',
-    },
-  ]
-  const rightNavItems = [
-    {
-      name: 'về chúng tôi',
-      href: '/about-us',
-    },
-    {
-      name: 'liên hệ',
-      href: '/contact',
-    },
-  ]
+export default function NavBar({
+  logo,
+  navItems,
+  cartItemCount,
+  textColor = 'white',
+  bgColor = 'transparent',
+  themeColor = '#82AF38',
+}: NavBarProps) {
+  const [activeDropdown, setActiveDropdown] = useState<
+    'cart' | 'account' | 'menu' | null
+  >(null)
+  const cartRef = useRef<HTMLDivElement>(null)
+  const accountRef = useRef<HTMLDivElement>(null)
+  const menuRef = useRef<HTMLDivElement>(null)
+  const [isMobile, setIsMobile] = useState(false)
 
-  if (isMobile) {
-    return <MobileNavigation />
+  // Check for mobile screen size
+  useEffect(() => {
+    const checkMobile = () => {
+      setIsMobile(window.innerWidth <= 600)
+    }
+    checkMobile()
+    window.addEventListener('resize', checkMobile)
+    return () => window.removeEventListener('resize', checkMobile)
+  }, [])
+
+  // Close dropdown when clicking outside
+  useEffect(() => {
+    function handleClickOutside(event: MouseEvent) {
+      const target = event.target as Element
+
+      // Check if clicking on dropdown buttons
+      const clickingDropdownButton = target.closest('[data-dropdown]')
+      if (clickingDropdownButton) {
+        const buttonType = clickingDropdownButton.getAttribute('data-dropdown')
+        // If clicking cart/account from menu, don't close menu
+        if (
+          activeDropdown === 'menu' &&
+          (buttonType === 'cart' || buttonType === 'account')
+        ) {
+          return
+        }
+      }
+
+      // Check each dropdown separately
+      if (
+        activeDropdown === 'cart' &&
+        cartRef.current &&
+        !cartRef.current.contains(target)
+      ) {
+        setActiveDropdown(null)
+      }
+
+      if (
+        activeDropdown === 'account' &&
+        accountRef.current &&
+        !accountRef.current.contains(target)
+      ) {
+        setActiveDropdown(null)
+      }
+
+      if (
+        activeDropdown === 'menu' &&
+        menuRef.current &&
+        !menuRef.current.contains(target)
+      ) {
+        setActiveDropdown(null)
+      }
+    }
+
+    document.addEventListener('mousedown', handleClickOutside)
+    return () => {
+      document.removeEventListener('mousedown', handleClickOutside)
+    }
+  }, [activeDropdown])
+
+  // Toggle dropdown visibility
+  const toggleDropdown = (dropdown: 'cart' | 'account' | 'menu') => {
+    setActiveDropdown((prev) => {
+      // If clicking the same dropdown that's already open, close it
+      if (prev === dropdown) return null
+
+      return dropdown
+    })
   }
 
   return (
-    <nav className='bg-primary text-white rounded-full px-8 py-4 max-w-3xl w-full mx-auto'>
-      <div className='flex items-center relative'>
-        {/* Left Navigation Items */}
-
-        <div className='flex items-center space-x-8'>
-          {leftNavItems.map((item) => (
-            <Link key={item.name} href={item.href}>
-              <button
-                className='uppercase font-semibold text-xs tracking-wider hover:text-gold duration-300 cursor-pointer'
-                data-testid='button-trending'
-              >
-                {item.name}
-              </button>
-            </Link>
-          ))}
-        </div>
-
-        {/* Centered Circular Logo */}
-        <div className='absolute left-1/2 transform -translate-x-1/2 -translate-y-1'>
-          <Link href='/'>
-            <div
-              className='w-20 h-20 bg-primary rounded-full flex flex-col items-center justify-center cursor-pointer transition-transform duration-300'
-              data-testid='button-logo'
-            >
-              <Image
-                alt='Santorini logo'
-                src='/svg/logo.svg'
-                width={60}
-                height={60}
-              />
-            </div>
-          </Link>
-        </div>
-        <div className='w-32' />
-        {/* Right Navigation Items */}
-        <div className='flex items-center space-x-8'>
-          {rightNavItems.map((item) => (
-            <Link key={item.name} href={item.href}>
-              <button
-                className='uppercase font-semibold text-xs tracking-wider hover:text-gold duration-300 cursor-pointer'
-                data-testid='button-trending'
-              >
-                {item.name}
-              </button>
-            </Link>
-          ))}
-          <div className='flex items-center space-x-6'>
-            <Badge
-              color='secondary'
-              content={5}
-              shape='rectangle'
-              size='sm'
-              showOutline={false}
-              variant='shadow'
-            >
-              <ShoppingCart className='cursor-pointer' size={18} />
-            </Badge>
-            <UserRound className='cursor-pointer' size={18} />
-          </div>
-        </div>
+    <div
+      className='flex w-full max-w-[1440px] items-center justify-between pt-6 ps-[6.2%] pe-[5.5%] py-3 z-[999] select-none mx-auto'
+      style={{
+        backgroundColor: bgColor,
+        backdropFilter: 'blur(2px)',
+        transition: 'background-color 0.3s ease',
+      }}
+    >
+      {/* Logo */}
+      <div
+        className='text-2xl uppercase font-bold cursor-default'
+        style={{
+          color: textColor,
+          fontWeight: 900,
+        }}
+      >
+        {logo}
       </div>
-    </nav>
-  )
-}
 
-const MobileNavigation = () => {
-  return (
-    <div className='fixed bottom-4 left-4 right-4 z-50'>
-      <div className='bg-coffee-dark rounded-2xl p-4 shadow-lg'>
-        <div className='flex justify-around items-center'>
-          <Link href='/trending'>
-            <button
-              className='flex flex-col items-center space-y-1 transition-colors duration-300 hover:text-gold'
-              data-testid='button-mobile-trending'
-            >
-              <Flame className='text-gold' size={20} />
-              <span className='text-xs'>HOME</span>
-            </button>
-          </Link>
-          <Link href='/rewards'>
-            <button
-              className='flex flex-col items-center space-y-1 transition-colors duration-300 hover:text-gold'
-              data-testid='button-mobile-rewards'
-            >
-              <Gift className='text-gold' size={20} />
-              <span className='text-xs'>MENU</span>
-            </button>
-          </Link>
-          <Link href='/'>
-            <button
-              className='flex flex-col items-center space-y-1 transition-colors duration-300 hover:text-gold'
-              data-testid='button-mobile-shop'
-            >
-              <Coffee className='text-gold' size={20} />
-              <span className='text-xs'>REWARDS</span>
-            </button>
-          </Link>
-          <Link href='/reserve'>
-            <button
-              className='flex flex-col items-center space-y-1 transition-colors duration-300 hover:text-gold'
-              data-testid='button-mobile-reserve'
-            >
-              <Calendar className='text-gold' size={20} />
-              <span className='text-xs'>ABOUT US</span>
-            </button>
-          </Link>
-          <Link href='/profile'>
-            <button
-              className='flex flex-col items-center space-y-1 transition-colors duration-300 hover:text-gold'
-              data-testid='button-mobile-profile'
-            >
-              <User className='text-gold' size={20} />
-              <span className='text-xs'>CONTACT US</span>
-            </button>
-          </Link>
+      {/* Navigation Links - Desktop */}
+      <div className='hidden space-x-8 md:flex xl:space-x-20'>
+        {navItems.map((item, index) => (
+          <span
+            key={index}
+            className='text-base font-normal cursor-pointer relative group'
+            style={{ color: textColor }}
+          >
+            {item.label}
+            <span
+              className='absolute left-0 bottom-0 w-0 h-[2px] group-hover:w-full transition-all duration-300'
+              style={{ backgroundColor: textColor }}
+            />
+          </span>
+        ))}
+      </div>
+
+      {/* Mobile Menu */}
+      {isMobile && (
+        <div ref={menuRef}>
+          <MobileMenu
+            navItems={navItems}
+            cartItemCount={cartItemCount}
+            themeColor={themeColor}
+            textColor={textColor}
+            activeDropdown={activeDropdown}
+            toggleDropdown={toggleDropdown}
+          />
         </div>
+      )}
+
+      {/* User and Cart Icons - Desktop */}
+      <div className={`${isMobile ? 'hidden' : 'flex'} items-center space-x-4`}>
+        {/* Account Icon with Dropdown */}
+        <div ref={accountRef} className='relative'>
+          <span
+            className='cursor-pointer hover:opacity-80 transition-opacity duration-300'
+            style={{ color: textColor }}
+            onClick={() => toggleDropdown('account')}
+          >
+            <User size={20} />
+          </span>
+
+          {/* Account Dropdown Menu - Desktop */}
+          {!isMobile && activeDropdown === 'account' && (
+            <AccountDropdown themeColor={themeColor} isDesktop={true} />
+          )}
+        </div>
+        <LocaleSwitcher />
       </div>
     </div>
   )
 }
-
-export default NavBar
