+++ src/i18n/navigation.ts
@@ -1,0 +1,7 @@
+import {createNavigation} from 'next-intl/navigation';
+
+import { routing } from '@/i18n/routing';
+
+
+export const {Link, getPathname, redirect, usePathname, useRouter} =
+  createNavigation(routing);
