+++ src/i18n/routing.ts
@@ -1,0 +1,10 @@
+import {defineRouting} from 'next-intl/routing';
+ 
+export const routing = defineRouting({
+  // A list of all locales that are supported
+  locales: ['en', 'vi'],
+ 
+  // Used when no locale matches
+  defaultLocale: 'vi',
+  localePrefix: "as-needed"
+});
