+++ src/middleware.ts
@@ -1,9 +1,15 @@
 import { type NextRequest } from 'next/server'
+import createMiddleware from 'next-intl/middleware';
 
 import { updateSession } from '@/lib/supabase/middleware'
 
+import { routing } from '@/i18n/routing'
+
+const handleI18nRouting = createMiddleware(routing)
+
 export async function middleware(request: NextRequest) {
-  return await updateSession(request)
+  const response = handleI18nRouting(request)
+  return await updateSession(request, response)
 }
 
 export const config = {
@@ -17,5 +23,6 @@ export const config = {
      * Feel free to modify this pattern to include more paths.
      */
     '/((?!_next/static|_next/image|favicon.ico|.*\\.(?:svg|png|jpg|jpeg|gif|webp)$).*)',
+    '/', '/(vi|en)/:path*'
   ],
 }
